Moved data files viewset into its own file.
