+++ mathesar/urls.py
@@ -3,6 +3,7 @@ from rest_framework_nested import routers
 
 from mathesar import views
 from mathesar.api import legacy_viewsets
+from mathesar.api.viewsets.data_files import DataFileViewSet
 from mathesar.api.viewsets.schemas import SchemaViewSet
 from mathesar.api.viewsets.tables import TableViewSet
 
@@ -11,7 +12,7 @@ router = routers.DefaultRouter()
 router.register(r'tables', TableViewSet, basename='table')
 router.register(r'schemas', SchemaViewSet, basename='schema')
 router.register(r'databases', legacy_viewsets.DatabaseViewSet, basename='database')
-router.register(r'data_files', legacy_viewsets.DataFileViewSet, basename='data-file')
+router.register(r'data_files', DataFileViewSet, basename='data-file')
 
 table_router = routers.NestedSimpleRouter(router, r'tables', lookup='table')
 table_router.register(r'records', legacy_viewsets.RecordViewSet, basename='table-record')
